pyproject.toml
[project]
name = "news_agent"
version = "1.0.0"
description = "An AI news agent demo"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.9",
    "langchain-core>=0.3.59",
    "langchain-openai",
    "langgraph>=0.4.2",
    "langsmith[pytest]>=0.3.4",
    "langgraph-cli[inmem]",
    "tavily-python",
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["news_agent"]

[tool.setuptools.package-dir]
"news_agent" = "src/news_agent"

[tool.setuptools.package-data]
"*" = ["py.typed"]
